test: 입력 예외 테스트를 다른 파일로 분리
